refactor: break out user list and other cmps from App.tsx
