+++ packages/app/src/App.tsx
@@ -1,124 +1,17 @@
-import React, { useRef, useState } from 'react';
-import { BrowserRouter, Route, Switch } from 'react-router-dom';
-import { QueryClient, QueryClientProvider, useMutation, useQuery } from 'react-query';
+import React from 'react';
+import { BrowserRouter, Link, Route, Switch } from 'react-router-dom';
+import { QueryClient, QueryClientProvider, useQuery } from 'react-query';
 import { ReactQueryDevtools } from 'react-query/devtools';
-import { makeGraphQLQuery, makeGraphQLFileUpload } from './dataservice';
+import { makeGraphQLQuery } from './dataservice';
 import secureLogo from './assets/secure.svg';
 import innovativeLogo from './assets/innovative.svg';
-import { Button } from './components/common';
-import Footer from './components/Footer';
-import Nav from './components/Nav';
 import LoginPage from './containers/LoginPage';
 import UserRegisterPage from './containers/UserRegisterPage';
+import UserListPage from './containers/UserListPage';
+import Spinner from './components/Spinner';
+import Layout from './containers/Layout';
 
 const queryClient = new QueryClient();
-
-// #region User Card *******
-
-type UserCardProps = {
-    user: {
-        id?: string;
-        displayName?: string;
-        email?: string;
-    };
-};
-
-function UserCard({ user }: UserCardProps): JSX.Element {
-    return (
-        <div className="flex p-4 transition duration-200 ease-in-out border-2 border-purple-600 rounded-md shadow-md transform hover:bg-purple-100 hover:scale-105 hover:shadow-lg">
-            <img
-                src="http://localhost:3000/default/avatar.png"
-                alt="default avatar"
-                className="w-24 rounded-full border-2 border-purple-300"
-            />
-            <div className="flex flex-col justify-center ml-4 text-purple-900">
-                <span>{user.displayName}</span>
-                <span>{user.email}</span>
-            </div>
-        </div>
-    );
-}
-
-// #endregion
-
-// #region User List *******
-
-type UserListProps = { className: string };
-
-function UserList({ className }: UserListProps): JSX.Element {
-    const statusQuery = 'query { users { id displayName email } }';
-    const { data, isLoading } = useQuery(['users', { query: statusQuery }], makeGraphQLQuery);
-    if (isLoading) return <Spinner />;
-    return (
-        <div className={className}>
-            <h2 className="text-3xl text-gray-600 text-center border-b-2">User List</h2>
-
-            <div className="px-8 py-4 md:p-4">
-                {data?.data?.users?.length > 0 ? (
-                    <div className="flex text-gray-500 text-lg">
-                        <span className="w-1/2 px-2">Id</span>
-                        <span className="w-1/2 px-2">Name</span>
-                        <span className="w-1/2 px-2">Email</span>
-                    </div>
-                ) : (
-                    <p className="mt-4 text-gray-800"> No users currently registered </p>
-                )}
-
-                {data &&
-                    data.data?.users?.map(
-                        (user: { id: string; displayName: string; email: string }) => {
-                            return (
-                                <div key={user.id} className="flex mb-2">
-                                    <span className="w-1/2 px-2 text-sm text-purple-900 md:text-base">
-                                        {user.id}
-                                    </span>
-                                    <span className="w-1/2 px-2 text-sm text-gray-800 md:text-base">
-                                        {user.displayName}
-                                    </span>
-                                    <span className="w-1/2 px-2 text-sm text-gray-800 md:text-base">
-                                        {user.email}
-                                    </span>
-                                </div>
-                            );
-                        }
-                    )}
-            </div>
-            <h2 className="mb-2 text-3xl text-gray-600 text-center">User Cards</h2>
-            <div className="grid grid-flow-rows grid-col-1 mx-8 lg:grid-cols-2 xl:grid-cols-3 gap-4">
-                {data &&
-                    data.data?.users?.map(
-                        (user: { id: string; displayName: string; email: string }) => {
-                            return <UserCard key={user.id} user={user} />;
-                        }
-                    )}
-            </div>
-        </div>
-    );
-}
-
-// #endregion
-
-// #region Spinner *******
-
-function Spinner(): JSX.Element {
-    return (
-        <div className="fixed top-0 right-0 h-screen w-screen z-50 flex justify-center items-center">
-            <div className="animate-spin rounded-full h-32 w-32 border-t-2 border-b-2 border-purple-900"></div>
-        </div>
-    );
-}
-
-/*
-function SmallSpinner(): JSX.Element {
-    return (
-        <div className="relative z-50 flex justify-center items-center">
-            <div className="animate-spin rounded-full h-16 w-16 border-t-2 border-b-2 border-purple-900"></div>
-        </div>
-    );
-}
-*/
-
-// #endregion
 
 // #region Hero Banner *******
 
@@ -240,168 +133,6 @@ function Features(): JSX.Element {
 
 // #endregion
 
-// #region File upload *******
-
-function FileUpload(): JSX.Element {
-    const imgRef = useRef<HTMLImageElement>(null);
-    const [highlight, setHighlight] = useState(false);
-    const [image, setImage] = useState<File | null>(null);
-    const [error, setError] = useState('');
-
-    const mutation = useMutation(makeGraphQLFileUpload);
-
-    const handleDragEnter: React.DragEventHandler<HTMLDivElement> = (e) => {
-        setHighlight(true);
-    };
-    const handleDragLeave: React.DragEventHandler<HTMLDivElement> = (e) => {
-        setHighlight(false);
-    };
-    const handleDragOver: React.DragEventHandler<HTMLDivElement> = (e) => {
-        e.preventDefault();
-    };
-
-    const handleDrop: React.DragEventHandler<HTMLDivElement> = (e) => {
-        e.preventDefault();
-        setHighlight(false);
-        const file = e.dataTransfer.files[0];
-        if (file) {
-            processFile(file);
-        }
-    };
-
-    const handleChooseFile: React.ChangeEventHandler<HTMLInputElement> = (e) => {
-        if (e.target && e.target.files) {
-            processFile(e.target.files[0]);
-        }
-    };
-
-    const handleFileUpload = (e: React.MouseEvent) => {
-        const AvatarUploadOp = `
-mutation AvatarUpload ($image: Upload!) {
-    avatarUpload(image: $image)
-}
-`;
-        const variables = { image: null, operationName: 'UploadAvatar' };
-
-        mutation.mutate(
-            { query: AvatarUploadOp, variables, file: image as File },
-            {
-                onSuccess: () => {
-                    clearFile();
-                }
-            }
-        );
-
-        // update to use the useMutation hook from react-query
-        // makeGraphQLFileUpload({ query: AvatarUploadOp, variables }, image as File).then(
-        //     ({ data }) => {
-        //         if (data.avatarUpload) {
-        //             clearFile();
-        //         }
-        //     }
-        // );
-    };
-
-    const processFile = (file: File): void => {
-        if (file.type === 'image/png') {
-            setImage(file);
-            setError('');
-            const reader = new FileReader();
-            reader.onloadend = function () {
-                if (imgRef.current) {
-                    imgRef.current.src = reader.result as string;
-                }
-            };
-            if (file) {
-                reader.readAsDataURL(file);
-            } else {
-                if (imgRef.current) {
-                    imgRef.current.src = '';
-                }
-            }
-        } else {
-            setError('Invalid file extension');
-        }
-    };
-
-    const clearFile = (): void => {
-        setImage(null);
-    };
-
-    return (
-        <div className="text-gray-700">
-            <h3 className="text-center text-2xl mb-2">Upload Image</h3>
-            <p className="italic text-center mb-2">
-                Note: Currently, the API drops this on the floor...
-            </p>
-            {!image && (
-                <div>
-                    <div
-                        className={`p-6 border-2 border-dashed rounded-lg text-center ${
-                            highlight ? 'border-green-600 bg-green-100' : 'border-gray-400'
-                        }`}
-                        onDragEnter={handleDragEnter}
-                        onDragLeave={handleDragLeave}
-                        onDragOver={handleDragOver}
-                        onDrop={handleDrop}
-                    >
-                        Drop Zone <em>(.png only)</em>
-                    </div>
-                    {error && <p className="mt-2 text-red-700">Error: {error}</p>}
-                    <div className="mt-8">
-                        <label
-                            htmlFor="file-upload"
-                            className="border-2 rounded-lg p-4 cursor-pointer"
-                        >
-                            Choose File
-                        </label>
-                        <input
-                            type="file"
-                            id="file-upload"
-                            className="hidden"
-                            accept="image/png"
-                            onChange={handleChooseFile}
-                        />
-                    </div>
-                </div>
-            )}
-            {image && (
-                <div>
-                    <div className="mt-4 p-4 flex border-2 rounded-lg">
-                        <img
-                            src=""
-                            className="h-24 w-24 rounded-full"
-                            alt="Avatar preview"
-                            ref={imgRef}
-                        />
-                        <div className="ml-8 pt-4 text-gray-700 flex justify-between w-full">
-                            <span className="">{image.name}</span>
-                            <span>{image.size / 1000} Kb</span>
-                        </div>
-                    </div>
-                    <div className="flex justify-end">
-                        <Button className="my-4 mr-6" variant="secondary" clickAction={clearFile}>
-                            Cancel
-                        </Button>
-                        <Button
-                            type="button"
-                            className="my-4"
-                            variant="primary"
-                            clickAction={handleFileUpload}
-                        >
-                            Upload
-                        </Button>
-                    </div>
-                </div>
-            )}
-        </div>
-    );
-}
-
-// #endregion
-
-// #region Home *******
-
 function Home(): JSX.Element {
     return (
         <>
@@ -414,32 +145,13 @@ function Home(): JSX.Element {
             </div>
 
             <Features />
-            <UserList className="w-full mx-auto mt-12 py-6 md:w-3/4 md:py-0" />
 
-            <div className="my-4 mx-auto max-w-md">
-                <FileUpload />
+            <div className="mt-12 max-w-md mx-auto text-lg text-gray-600 border-2 border-gray-600 rounded-md py-2 text-center">
+                <Link to="/users">Show Users</Link>
             </div>
         </>
     );
 }
-
-// #endregion
-
-// #region Layout *******
-
-function Layout({ children }: { children: React.ReactNode }): JSX.Element {
-    return (
-        <div className="flex flex-col h-screen">
-            <Nav />
-            <div className="flex-grow flex-shrink-0 pb-6">{children}</div>
-            <Footer />
-        </div>
-    );
-}
-
-// #endregion
-
-// #region App *******
 
 function App(): JSX.Element {
     return (
@@ -456,6 +168,9 @@ function App(): JSX.Element {
                         <Route exact path="/register">
                             <UserRegisterPage />
                         </Route>
+                        <Route exact path="/users">
+                            <UserListPage />
+                        </Route>
                     </Switch>
                     <ReactQueryDevtools initialIsOpen={false} />
                 </QueryClientProvider>
@@ -463,7 +178,5 @@ function App(): JSX.Element {
         </BrowserRouter>
     );
 }
-
-// #endregion
 
 export default App;
