+++ packages/app/src/components/Spinner/index.tsx
@@ -1,0 +1,21 @@
+import React from 'react';
+
+function Spinner(): JSX.Element {
+    return (
+        <div className="fixed top-0 right-0 h-screen w-screen z-50 flex justify-center items-center">
+            <div className="animate-spin rounded-full h-32 w-32 border-t-2 border-b-2 border-purple-900"></div>
+        </div>
+    );
+}
+
+export default Spinner;
+
+/*
+function SmallSpinner(): JSX.Element {
+    return (
+        <div className="relative z-50 flex justify-center items-center">
+            <div className="animate-spin rounded-full h-16 w-16 border-t-2 border-b-2 border-purple-900"></div>
+        </div>
+    );
+}
+*/
